+++ src/pages/addCard/hooks/useCardNumberInput.ts
@@ -1,7 +1,7 @@
 import { useCallback } from "react";
 import { insertAtInterval, replaceTo } from "../../../utils/stringUtils";
 import useDisplayedInput from "../../../hooks/useDisplayedInput";
-import { isNumber } from "../../../utils/validation";
+import { isNumber } from "../../../utils/validationUtils";
 
 const CARN_NUMBER_LENGTH = 16;
 const CARD_NUMBER_HYPHEN_INTERVAL = 4;
