+++ src/utils/validationUtils.ts
